requirements.txt
Django==1.6.2
South==0.8.4
django-cors-headers==1.0.0
django-extensions==1.5.7
django-filter==0.8
django-recaptcha==1.0
mock==1.3.0
newrelic==2.22.1.20
psycopg2==2.5.3
Pillow==3.0.0
python-dateutil==2.2
requests==2.3.0
simplejson==3.6.5
uWSGI==2.0.5.1
wsgiref==0.1.2